requirements.txt
aiogram>=3.0.0
loguru
python-dotenv
